requirements.txt
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
execnet==1.9.0
executing==0.8.3
fastjsonschema==2.15.3
feature-engine==1.4.1
filelock==3.9.0
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lightgbm==3.3.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numpy==1.22.3
packaging==22.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.6.2
plotly==5.10.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pycparser==2.21
pydantic==1.10.2
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyproject_api==1.4.0
pyrsistent==0.18.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
regex==2022.8.17
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.3
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.12.0
strictyaml==1.6.2
tenacity==8.0.1
terminado==0.13.3
threadpoolctl==3.1.0
tid-costa-rica-poverty-prediction==0.0.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tox==4.2.3
tqdm==4.64.0
traitlets==5.2.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
virtualenv==20.17.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wordcloud==1.8.2.2
zipp==3.8.1
